+++ typescript/sdk/src/utils.ts
@@ -1,6 +1,5 @@
 import { arrayify, BytesLike, hexlify } from '@ethersproject/bytes';
 import { ethers } from 'ethers';
-import { BigNumber, FixedNumber } from 'ethers';
 
 export type Address = string;
 
@@ -52,44 +51,4 @@ export function evmId(data: BytesLike): Address {
  */
 export function delay(ms: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-/**
- * Converts a BigNumber to a FixedNumber of the format fixed128x18.
- * @param big The BigNumber to convert.
- * @returns A FixedNumber representation of a BigNumber.
- */
-export function bigToFixed(big: BigNumber): FixedNumber {
-  return FixedNumber.from(big.toString());
-}
-
-/**
- * Converts a FixedNumber (of any format) to a BigNumber.
- * @param fixed The FixedNumber to convert.
- * @param ceil If true, the ceiling of fixed is used. Otherwise, the floor is used.
- * @returns A BigNumber representation of a FixedNumber.
- */
-export function fixedToBig(
-  fixed: FixedNumber,
-  ceil: boolean = false,
-): BigNumber {
-  const fixedAsInteger = ceil ? fixed.ceiling() : fixed.floor();
-  return BigNumber.from(fixedAsInteger.toFormat('fixed256x0').toString());
-}
-
-/**
- * Multiplies a BigNumber by a FixedNumber, returning the BigNumber product.
- * @param big The BigNumber to multiply.
- * @param fixed The FixedNumber to multiply.
- * @param ceil If true, the ceiling of the product is used. Otherwise, the floor is used.
- * @returns The BigNumber product.
- */
-export function mulBigAndFixed(
-  big: BigNumber,
-  fixed: FixedNumber,
-  ceil: boolean = false,
-): BigNumber {
-  // Converts big to a FixedNumber, multiplies it by fixed, and converts the product back
-  // to a BigNumber.
-  return fixedToBig(fixed.mulUnsafe(bigToFixed(big)), ceil);
 }
